requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.4
kafka-python-ng==2.2.2
MarkupSafe==2.1.5
packaging==24.0
pyparsing==3.1.2
requests==2.32.3
setuptools==70.3.0
soupsieve==2.6
urllib3==1.26.18
Werkzeug==3.0.3

